frontend/requirements.txt
streamlit
requests
sseclient-py
python-dotenv